Components placement refactorings
