+++ src/Homes/index.js
@@ -3,7 +3,7 @@ import { Route } from "react-router-dom";
 
 import Navbar from "../Navbar";
 import Footer from "../Footer";
-import Hidden from "../Hidden";
+import VisuallyHidden from "../UI/VisuallyHidden";
 import List from "./List";
 import Show from "./Show";
 
@@ -17,9 +17,9 @@ export default function() {
         <Route exact path="/homes/:id" component={Show} />
       </main>
 
-      <Hidden>
+      <VisuallyHidden>
         <Footer />
-      </Hidden>
+      </VisuallyHidden>
     </div>
   );
 }
