+++ src/Navbar/index.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 
 import bp from "../breakpoints";
 
-import Container from "../Container";
+import Container from "../UI/Container";
 import Menu from "./Menu";
 import Search from "./Search";
