+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import "normalize.css";
-import "flexboxgrid2";
+//import "flexboxgrid2";
 import "./fonts/fonts.css";
 import "./index.css";
